Dog_Cat_Classifier/requirement.txt
absl-py==0.8.1
alabaster==0.7.8
asn1crypto==0.24.0
astor==0.8.0
astroid==2.1.0
atomicwrites==1.1.5
attrs==18.2.0
Babel==2.6.0
beautifulsoup4==4.7.1
biopython==1.73
bleach==3.1.0
blosc==1.7.0
bottle==0.12.15
Brlapi==0.6.7
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
chrome-gnome-shell==0.0.0
cloudpickle==0.8.0
cmake==3.13.3
coverage==4.5.2
cryptography==2.6.1
cson==0.7
cupshelpers==1.0
cvxpy==1.0.25
cycler==0.10.0
Cython==0.29.2
dask==1.0.0+dfsg
decorator==4.3.0
defusedxml==0.5.0
dill==0.3.1.1
distro==1.3.0
distro-info==0.21
dlib==19.19.0
docutils==0.14
ecos==2.0.7.post1
entrypoints==0.3
extras==1.0.0
fixtures==3.0.0
frozendict==1.2
future==0.16.0
gast==0.3.2
GDAL==2.4.0
gnome-gmail==2.6
google-pasta==0.1.8
graphviz==0.8.4
grpcio==1.25.0
h5py==2.8.0
html5lib==1.0.1
idna==2.6
imageio==2.6.1
imagesize==1.0.0
iotop==0.6
ipdb==0.11
ipykernel==4.9.0
ipympl==0.3.3
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
jedi==0.13.2
Jinja2==2.10
joblib==0.13.0
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
linecache2==1.0.0
loader==2017.9.11
locket==0.2.0
logilab-common==1.4.2
louis==3.8.0
lxml==4.3.2
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
mnist==0.2.2
mock==2.0.0
more-itertools==4.2.0
mpi4py==2.0.0
mpmath==1.1.0
multiprocess==0.70.9
mxnet==1.5.1.post0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.4.5
nodejs==0.1.1
nose==1.3.7
nose2==0.8.0
notebook==5.7.8
npm==0.1.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.7.0
olefile==0.46
opencv-python==4.1.2.30
optional-django==0.1.0
osqp==0.6.1
packaging==19.0
pandas==0.23.3+dfsg
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.9
pathlib2==2.3.5
patsy==0.5.0+dev
pbr==4.2.0
pdfarranger==1.1.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.8.0
prometheus-client==0.6.0
prompt-toolkit==1.0.15
protobuf==3.10.0
psutil==5.5.1
py==1.7.0
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydot==1.4.1
pyFFTW==0.11.1
pyflakes==2.0.0
Pygments==2.3.1
PyGObject==3.30.4
pylint==2.2.2
pyparsing==2.2.0
PyPDF2==1.26.0
pyserial==3.4
pysmbc==1.0.15.6
pytest==3.10.1
python-apt==1.8.4
python-dateutil==2.7.3
python-igraph==0.7.1.post6
python-mimeparse==0.1.4
pytz==2019.1
PyWavelets==0.5.1
pyxattr==0.6.1
pyxdg==0.25
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
reportlab==3.5.13
requests==2.21.0
requests-futures==0.9.9
roman==2.0.0
rope==0.10.5
rpy2==2.9.5
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.1.0
scour==0.37
scs==2.1.1.post2
seaborn==0.9.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
sklearn==0.0
soupsieve==1.8
speg==0.3
Sphinx==1.8.4
spyder==3.3.3
spyder-kernels==0.2.4
sympy==1.4
tables==3.4.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
testtools==2.3.0
texttable==1.6.0
toolz==0.9.0
torch==1.3.1+cpu
torchsummary==1.5.1
torchvision==0.4.2+cpu
tornado==5.1.1
traceback2==1.4.0
traitlets==4.3.2
unattended-upgrades==0.1
unicodecsv==0.14.1
unittest2==1.1.0
urllib3==1.24.1
waitress==1.2.0b2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.5.0
wrapt==1.11.2
youtube-dl==2019.1.17
